feat: debugger detection added
* add DebuggerDetection.swift

* add DebuggerDetection to ThreatDetectionCenter

* add debugger detection view to SecurityToolkitExample

* add documentation for isDebuggerAttached

* fix typo: isDebuggerAttached -> isDebuggerDetected

* Comment made even more clear

* hasTracerFlagSet() additionally made private

* No explicit Darwin import needed and therefore rmd

no explicit Darwin import needed and therefore removed

* Use of optional bool for isDebuggerDetected return

Use of optional bool for public function isDebuggerDetected.

This decision is the result of a joint discussion that took place in the associated PR #10 :
https://github.com/EXXETA/iOS-Security-Toolkit/pull/10

Motivation:
No exception should be thrown.
It should be communicated that something went wrong in the decision-making process and that no clear statement can be made.

Decision:
We now return nil in an unclear case.
nil: The detection process did not produce a definitive result. This could happen due to system limitations, lack of required permissions, or other undefined conditions.